taxi/requirements.txt
Django==4.1.5
pre-commit==2.21.0
django-filter==22.1
djangorestframework==3.14.0
django-environ==0.9.0
psycopg2-binary==2.9.5
gunicorn==20.1.0
# Testing
pytest-django==4.5.2
coverage==7.0.2
pytest-mock==3.10.0
pytest-drf==1.1.3
model-bakery==1.9.0
# S3 storage
boto3==1.26.99
django-storages==1.13.2
numpy==1.24.2
requests==2.28.2
fideslang==1.4.1